Day 6 Commit #1
I have added a small amount of code that shows if the user is online or offline. A new field "status" is added to the realtime database which updates when the user becomes online or goes offline. An green color ImageView is shown if status == "online" and hidden when status == "offline".