Day 8 Commit #1
It has almost been a month since I have updated this project. Today, I have added the functionality of Reset Password where a password will be sent to the email ID of the user with a link. This link will redirect him/her to a firebase web page where their password can be reset.